+++ bin/download_chromium.dart
@@ -2,6 +2,6 @@ import 'package:puppeteer/puppeteer.dart';
 
 void main() async {
   var info = await downloadChrome();
-  print('Chromium ${info.revision} downloaded in ${info.folderPath}.'
+  print('Chromium ${info.version} downloaded in ${info.folderPath}.'
       '\nExecutable: ${info.executablePath}');
 }