+++ lib/src/downloader.dart
@@ -4,49 +4,51 @@ import 'package:archive/archive.dart';
 import 'package:http/http.dart' as http;
 import 'package:path/path.dart' as p;
 
-const int _lastRevision = 1108766;
+const _lastVersion = '113.0.5672.63';
 
-class RevisionInfo {
+class DownloadedBrowserInfo {
   final String executablePath;
   final String folderPath;
-  final int revision;
+  final String version;
 
-  RevisionInfo(
+  DownloadedBrowserInfo(
       {required this.executablePath,
       required this.folderPath,
-      required this.revision});
+      required this.version});
 }
 
 ///
 /// Downloads the chrome revision specified by [revision] to the [cachePath] directory.
 /// ```dart
 /// await downloadChrome(
-///   revision: 1083080,
-///   cachePath: '.local-chromium',
+///   version: '112.0.5615.121',
+///   cachePath: '.local-chrome',
 ///   onDownloadProgress: (received, total) {
 ///     print('downloaded $received of $total bytes');
 ///   });
 /// ```
-Future<RevisionInfo> downloadChrome({
-  int? revision,
+Future<DownloadedBrowserInfo> downloadChrome({
+  String? version,
   String? cachePath,
   void Function(int received, int total)? onDownloadProgress,
+  BrowserPlatform? platform,
 }) async {
-  revision ??= _lastRevision;
-  cachePath ??= '.local-chromium';
+  version ??= _lastVersion;
+  cachePath ??= '.local-chrome';
+  platform ??= BrowserPlatform.current;
 
-  var revisionDirectory = Directory(p.join(cachePath, '$revision'));
+  var revisionDirectory = Directory(p.join(cachePath, version));
   if (!revisionDirectory.existsSync()) {
     revisionDirectory.createSync(recursive: true);
   }
 
-  var exePath = getExecutablePath(revisionDirectory.path);
+  var exePath = p.join(revisionDirectory.path, getExecutablePath(platform));
 
   var executableFile = File(exePath);
 
   if (!executableFile.existsSync()) {
-    var url = _downloadUrl(revision);
-    var zipPath = p.join(cachePath, '${revision}_${p.url.basename(url)}');
+    var url = _downloadUrl(platform, version);
+    var zipPath = p.join(revisionDirectory.path, p.url.basename(url));
     await _downloadFile(url, zipPath, onDownloadProgress);
     _unzip(zipPath, revisionDirectory.path);
     File(zipPath).deleteSync();
@@ -66,10 +68,10 @@ Future<RevisionInfo> downloadChrome({
     await Process.run('xattr', ['-d', 'com.apple.quarantine', chromeAppPath]);
   }
 
-  return RevisionInfo(
+  return DownloadedBrowserInfo(
       folderPath: revisionDirectory.path,
       executablePath: executableFile.path,
-      revision: revision);
+      version: version);
 }
 
 Future _downloadFile(
@@ -100,7 +102,14 @@ void _unzip(String path, String targetPath) {
     // The _simpleUnzip doesn't support symlinks so we prefer a native command
     Process.runSync('unzip', [path, '-d', targetPath]);
   } else {
-    _simpleUnzip(path, targetPath);
+    try {
+      var result = Process.runSync('tar', ['-xf', path, '-C', targetPath]);
+      if (result.exitCode != 0) {
+        throw Exception('Failed to unzip chrome binaries:\n${result.stderr}');
+      }
+    } on ProcessException {
+      _simpleUnzip(path, targetPath);
+    }
   }
 }
 
@@ -126,30 +135,52 @@ void _simpleUnzip(String path, String targetPath) {
   }
 }
 
-const _baseUrl = 'https://storage.googleapis.com/chromium-browser-snapshots';
+const _baseUrl = 'https://edgedl.me.gvt1.com/edgedl/chrome/chrome-for-testing';
 
-String _downloadUrl(int revision) {
-  if (Platform.isWindows) {
-    return '$_baseUrl/Win_x64/$revision/chrome-win.zip';
-  } else if (Platform.isLinux) {
-    return '$_baseUrl/Linux_x64/$revision/chrome-linux.zip';
-  } else if (Platform.isMacOS) {
-    return '$_baseUrl/Mac/$revision/chrome-mac.zip';
-  } else {
-    throw UnsupportedError(
-        "Can't download chrome for platform ${Platform.operatingSystem}");
-  }
+String _downloadUrl(BrowserPlatform platform, String version) {
+  return '$_baseUrl/$version/${platform.folder}/chrome-${platform.folder}.zip';
 }
 
-String getExecutablePath(String revisionPath) {
-  if (Platform.isWindows) {
-    return p.join(revisionPath, 'chrome-win', 'chrome.exe');
-  } else if (Platform.isLinux) {
-    return p.join(revisionPath, 'chrome-linux', 'chrome');
-  } else if (Platform.isMacOS) {
-    return p.join(revisionPath, 'chrome-mac', 'Chromium.app', 'Contents',
-        'MacOS', 'Chromium');
-  } else {
-    throw UnsupportedError('Unknown platform ${Platform.operatingSystem}');
+String getExecutablePath(BrowserPlatform platform) {
+  return switch (platform) {
+    BrowserPlatform.macArm64 || BrowserPlatform.macX64 => p.join(
+        'chrome-${platform.folder}',
+        'Google Chrome for Testing.app',
+        'Contents',
+        'MacOS',
+        'Google Chrome for Testing'),
+    BrowserPlatform.linux64 => p.join('chrome-${platform.folder}', 'chrome'),
+    BrowserPlatform.windows32 || BrowserPlatform.windows64 =>
+      p.join('chrome-${platform.folder}', 'chrome.exe'),
+  };
+}
+
+enum BrowserPlatform {
+  macArm64._('macos_arm64', 'mac-arm64'),
+  macX64._('macos_x64', 'mac-x64'),
+  linux64._('linux_x64', 'linux64'),
+  windows32._('windows_ia32', 'win32'),
+  windows64._('windows_x64', 'win64'),
+  ;
+
+  final String dartPlatform;
+  final String folder;
+
+  const BrowserPlatform._(this.dartPlatform, this.folder);
+
+  factory BrowserPlatform.fromDartPlatform(String versionStringFull) {
+    final split = versionStringFull.split('"');
+    if (split.length < 2) {
+      throw FormatException(
+          "Unknown version from Platform.version '$versionStringFull'.");
+    }
+    final versionString = split[1];
+    return values.firstWhere((e) => e.dartPlatform == versionString,
+        orElse: () => throw FormatException(
+            "Unknown '$versionString' from Platform.version"
+            " '$versionStringFull'."));
   }
+
+  static final BrowserPlatform current =
+      BrowserPlatform.fromDartPlatform(Platform.version);
 }