+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 3.0.0
+- Download "Chrome for Testing" instead of Chromium.
+- Update to Chrome 113
+- Cache the browser binaries by default in `.local-chrome` instead of `.local-chromium`
+- Replace `int revision` parameter with `String version` in `downloadChrome`
+- Require minimum Dart `3.0.0` version.
+
 ## 2.24.0
 - Update to Chromium 112